pyproject.toml
[tool.poetry]
name = "hosterhase"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.6.0"
telethon = "^1.33.1"
python-dotenv = "^1.0.0"
promptflow = "^1.2.0"
promptflow-tools = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
